Add test for built CommonJS code.
