+++ tests/test-mocha.cjs
@@ -1,0 +1,2 @@
+const {should} = require('chai');
+global.should = should();
